@qawolf/web: separate query helpers to their own file
