+++ packages/browser/tests/web/find.test.ts
@@ -2,6 +2,7 @@ import { CONFIG } from "@qawolf/config";
 import { QAWolfWeb } from "@qawolf/web";
 import { Page } from "puppeteer";
 import { Browser } from "../../src/Browser";
+import { hasText } from "../../src/find";
 
 let browser: Browser;
 let page: Page;
@@ -41,138 +42,26 @@ describe("findClickableAncestor", () => {
   });
 });
 
-describe("queryActionElements", () => {
-  it("returns all elements for click action", async () => {
-    const actionElementCount = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const actionElements = qawolf.find.queryActionElements("click");
-
-      return actionElements.length;
-    });
-
-    expect(actionElementCount).toBe(28);
+describe("hasText", () => {
+  beforeAll(async () => {
+    browser = await Browser.create({ url: `${CONFIG.testUrl}login` });
+    page = await browser.currentPage();
   });
 
-  it("returns only input elements for input action", async () => {
-    const actionElements = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const actionElements = qawolf.find.queryActionElements("type");
+  afterAll(() => browser.close());
 
-      return actionElements.map((el: HTMLElement) => qawolf.xpath.getXpath(el));
-    });
-
-    expect(actionElements).toEqual([
-      "//*[@id='username']",
-      "//*[@id='password']"
-    ]);
+  it("returns true if page has text", async () => {
+    const result = await hasText(page, "tomsmith");
+    expect(result).toBe(true);
   });
 
-  it("does not include elements that are not visible", async () => {
-    const actionElementXpaths = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const username = document.getElementById("username")!;
-      username.style.border = "0";
-      username.style.padding = "0";
-      username.style.width = "0";
-
-      const actionElements = qawolf.find.queryActionElements("type");
-
-      return actionElements.map((el: HTMLElement) => qawolf.xpath.getXpath(el));
-    });
-
-    expect(actionElementXpaths).toEqual(["//*[@id='password']"]);
-
-    await browser.goto(`${CONFIG.testUrl}login`); // reset styles
-  });
-});
-
-describe("queryDataElements", () => {
-  it("returns all elements with given data value for click action", async () => {
-    const dataElementXpaths = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const submit = document.getElementsByTagName("button")[0]!;
-      submit.setAttribute("data-qa", "submit");
-
-      const dataElements = qawolf.find.queryDataElements({
-        action: "click",
-        dataAttribute: "data-qa",
-        dataValue: "submit"
-      });
-      submit.removeAttribute("data-qa");
-
-      return dataElements.map((el: HTMLElement) => qawolf.xpath.getXpath(el));
-    });
-
-    expect(dataElementXpaths).toEqual(["//*[@id='login']/button"]);
+  it("returns false if page does not have text", async () => {
+    const result = await hasText(page, "janedoe", 250);
+    expect(result).toBe(false);
   });
 
-  it("returns only input elements with given data value for input action", async () => {
-    const dataElementXpaths = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const username = document.getElementById("username")!;
-      const submit = document.getElementsByTagName("button")[0]!;
-      username.setAttribute("data-qa", "username");
-      submit.setAttribute("data-qa", "username");
-
-      const dataElements = qawolf.find.queryDataElements({
-        action: "type",
-        dataAttribute: "data-qa",
-        dataValue: "username"
-      });
-
-      username.removeAttribute("data-qa");
-      submit.removeAttribute("data-qa");
-
-      return dataElements.map((el: HTMLElement) => qawolf.xpath.getXpath(el));
-    });
-
-    expect(dataElementXpaths).toEqual(["//*[@id='username']"]);
-  });
-
-  it("does not include elements that are not visible", async () => {
-    const dataElementXpaths = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const username = document.getElementById("username")!;
-      const password = document.getElementById("password")!;
-
-      username.style.border = "0";
-      username.style.padding = "0";
-      username.style.width = "0";
-
-      username.setAttribute("data-qa", "username");
-      password.setAttribute("data-qa", "username");
-
-      const dataElements = qawolf.find.queryDataElements({
-        action: "type",
-        dataAttribute: "data-qa",
-        dataValue: "username"
-      });
-
-      return dataElements.map((el: HTMLElement) => qawolf.xpath.getXpath(el));
-    });
-
-    expect(dataElementXpaths).toEqual(["//*[@id='password']"]);
-
-    await browser.goto(`${CONFIG.testUrl}login`); // reset styles/data
-  });
-});
-
-describe("queryVisibleElements", () => {
-  it("returns only visible elements for a given selector", async () => {
-    const visibleElementXpaths = await page.evaluate(() => {
-      const qawolf: QAWolfWeb = (window as any).qawolf;
-      const username = document.getElementById("username")!;
-      username.style.border = "0";
-      username.style.padding = "0";
-      username.style.width = "0";
-
-      const actionElements = qawolf.find.queryVisibleElements("input");
-
-      return actionElements.map((el: HTMLElement) => qawolf.xpath.getXpath(el));
-    });
-
-    expect(visibleElementXpaths).toEqual(["//*[@id='password']"]);
-
-    await browser.goto(`${CONFIG.testUrl}login`); // reset styles
+  it("returns false if timeout reached", async () => {
+    const result = await hasText(page, "sup", 0);
+    expect(result).toBe(false);
   });
 });
